requirements.txt
pytest~=8.3.3
selene~=2.0.0rc9
selenium~=4.26.1
allure-pytest
allure-python-commons~=2.13.5
utils~=1.0.2
python-dotenv




